+++ src/index.js
@@ -1,8 +1,8 @@
-import './style.css';
+import './css/style.css';
 import './css/home.css';
 import './css/canvas.css';
 import './assets/idea-logo-small.png';
-import router from './router';
+import router from './helpers/router';
 
 window.addEventListener('hashchange', router);
 window.addEventListener('load', router);
